LayoutTests/fast/dom/Element/body-scrollTop.html

<!DOCTYPE html>
<html>
    <head>
        <style>
            div {
                height: 9999px;
                width:9999px;
            }
        </style>
        <script src="../../../resources/js-test-pre.js"></script>
        <script>
            function runTest() {
                description('Tests that for non-standard mode document.body.scrollTop returns the scroll top value as 0');

                document.body.scrollTop = 500;
                shouldBe("document.body.scrollTop","500");
				isSuccessfullyParsed();
            }
        </script>
    </head>
    <body onload="runTest()">
        <div></div>
        <script src="../../../resources/js-test-post.js"></script>
    </body>
</html>
