LayoutTests/fast/dom/Element/body-scrollTop-expected.txt

Tests that for non-standard mode document.body.scrollTop returns the scroll top value as 0

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS document.body.scrollTop is 0
PASS successfullyParsed is true

TEST COMPLETE
